add password authentication support
